requirements.txt
anyio==4.12.0
certifi==2025.11.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
notion-client==2.7.0
python-dotenv==1.2.1
pytz==2025.2
schedule==1.2.2
typing_extensions==4.15.0
